requirements.txt
annotated-types==0.6.0
atlassian-python-api==3.41.11
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
Deprecated==1.2.14
idna==3.6
importlib_metadata==7.0.2
jmespath==1.0.1
oauthlib==3.2.2
pydantic==2.6.4
pydantic_core==2.16.3
pydantic_yaml==1.2.1
requests==2.31.0
requests-oauthlib==1.4.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
six==1.16.0
soupsieve==2.5
typing_extensions==4.10.0
urllib3==2.2.1
wrapt==1.16.0
zipp==3.17.0
